rename Login stylesheet for use with Reg
